jhursshop/requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
Django==4.1.7
django-cors-headers==3.14.0
django-defender==0.9.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.25.1
inflection==0.5.1
jsonschema==4.17.3
Pillow==9.4.0
PyJWT==2.6.0
pyrsistent==0.19.3
python-dotenv==1.0.0
pytz==2022.7.1
PyYAML==6.0
redis==4.5.4
sqlparse==0.4.3
tzdata==2022.7
uritemplate==4.1.1
whitenoise==6.4.0
